TuringPlayer rendering
changed track length to 1024.

changed iterator so that setting a value to zero doesn't actually allocate any space (it can be represented just as well by a virtual zero).
fixed a bug for iterator subtraction that caused infinite recursion (used - instead of +).

created a Rule class that holds rule data.
changed AddRule to take a Rule object instead.
added a method to remove a rule.
added a method to get the rule that is about to be executed.

moved dispose code from TuringPlayer into the user-defined file.
implemented the rendering features.