TuringPlayer execution + paging
added an AddRule overload for Turing that takes a rule object.

changed the rendering logic for TuringPlayer.
fixed a bug where it was outlining the cursor position even if it was out of range.
changed the renderer to expand to fill the form vertically (was previously just horizontal).

added a method to TuringPlayer to advance the display by a number of pages.
added FocusCursor() to advance to the page where the execution cursor is.
added buttons to page around the turing machine's memory.

added a Tick() method to TuringPlayer to tick the turing machine and re-render the display.
added a checkbox and a wrapper Follow that makes TuringPlayer.Tick() follow the execution cursor.