file io + performance + bug fixes
changed DataIterator to a struct (was initially planned to have additional info pertaining to a total data size limit, but that's been scrapped now).
much higher performance since it fits in a 64-bit register.

added ++ and -- operators for DataIterator (was hesitant to do them since C# gives no way to overload prefix and postfix separately: use with caution, it always acts as postfix).

changed Pos and State in Turing class to variables rather than accessors (Pos required it to have its data be assignable).

fixed a bug where checking the Follow box wouldn't focus on the cursor (was assigning directly rather than calling the wrapper property - this functionality is now in the CheckChanged event).
fixed a bug where the "Tick" field in the display was displaying the state instead.

moved initialization of render iterator/pos/ticks to Init().

added TuringExt.cs that will hold extra turing utilities that aren't strictly required for use of a Turing object (currently contains file io methods).
added a menu strip to TuringPlayer with a File dropdown for opening rules or data files.